Fix an issue caused by the reconfiguration of a CAN connection
When a CAN connection class is reconfigured (e.g. goes from using can0
to using can1 interface), if there is a _monitor_listener setup already,
the connection between the bus (or more specifically, the notifier) and
the _monitor_listener will get lost during the reconfiguration process.
This commit introduces a fix for that issue - always check if the
listener is attached to the current notifier.